Optionally set user execution policy
This part is optional, but it seems like a common problem.

After installing pm2, invoking it on the command line on a machine with an ExecutionPolicy set to Restricted will fail, even if you run Unblock-File beforehand.

This script checks the current user's ExecutionPolicy and, if it's Restricted or Undefined, offers to change it to RemoteSigned.