Compress npm cache for *nix bundles
After populating the npm cache, compress it into a gzipped tarball.

Before setup, delete any existing npm cache and decompress the bundle.

This solves the issue where the cache may not contain everything from the bundle if an `npm install` was run between the bundle and setup.

Note: As of now, there is no reason to run `npm install` ever.

This applies to the *nix installer only for now.