Improve Windows bundle / offline install speed
PowerShell's Compress-Archive and Expand-Archive commands can be alarmingly slow in some situations.

Now, both `bundle` and `setup` scripts use Get-Command to check for the presence of tar.exe, which Microsoft has started shipping in some builds of Windows.

If tar.exe is installed, the setup script will attempt to decompress the offline archive using it. If not, it falls back to using Expand-Archive.

The `bundle` command archives the cache using both Compress-Archive and tar.exe, so the `setup` script on the installation target machine has a zip file to fall back to if necessary.