Improve windows setup message
It is very important not to invoke pm2 immediately after running `npm run setup`. If you do so without first either re-loading the shell's environmental variables or just opening a new window, the old PM_HOME value will cause pm2 to attempt to spawn a new process.

Move the 'setup complete' message so that this disclaimer is last, for emphasis.