Remove pm2-windows-service
Write a new Windows Service that leverages the most recent version of node-windows instead.

pm2-windows-service was great, but it's completely unmaintained. There are a few promising forks, but none seem to be working correctly with the current release of Node right now.

This new service has no dependencies, and only requires `node-windows` to install / uninstall it.

This also involved revamping the install / remove scripts