Ability to change the PM2 Service User
I would like to be able to change the service user from the session environment variable. 

We sometimes need to use a domain user as an app may need access to files / folders on the network not just the local machine

The code is effectively using coalesce to return the PM2_SERVICE_ACCOUNT and password from $Env and if it's not available, it will default to the LocalService user.